refactor: extract settings section navigation
